melhores o sistema de leitura do array deixei ele fragmentando como o seu metodo separado para tem melhor forma de arruma e fazer test so site e ja arruma maior parte sele como na parte de select add div para add o metodo de scroll infinity para deixa otimizado e ainda mais pesquisa para nao deixa o usuario fica rolando mil itens mais ainda falta alguns retoques para deixa melhor

+++ index.php
@@ -103,6 +103,11 @@ if ($uri === "/calendario") {
     }
     exit;
 }
+
+if ($uri === "/buscaList") {
+    completeList::buscaDadosList()();
+    exit;
+}
 if ($uri === "/deleted") {
     Delete::delete();
     exit;
@@ -154,15 +159,7 @@ if ($uri === '/menssagem') {
     }
     exit;
 }
-// if ($uri === '/confirma') {
-//     AuthLogin::check();
-//     if ($isAjax) {
-//         require __DIR__ . "/app/view/vendor/tabelas/menuPainel/confirmeReivindica.php";
-//     } else {
-//         require $HomeGenciador;
-//     }
-//     exit;
-// }
+
 if ($uri === '/confirmaReivindica' && $_SERVER['REQUEST_METHOD'] === 'POST') {
     AuthLogin::check();
     revindicar::ConfirmoRevidicacao();
